# back-end/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Rest API for Full Stack challenge Ducks Feeding.",
    "version": "1.0.0",
    "title": "animal-feeding-service",
    "contact": { "email": "contact@matheusicaro.com" },
    "license": {
      "name": "MIT",
      "url": "https://github.com/matheusicaro/challenge-full-stack-ducks-feeding/blob/master/LICENSE"
    }
  },
  "host": "*",
  "basePath": "/api/v1",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/user/singup": {
      "post": {
        "tags": ["User"],
        "summary": "Register new user",
        "description": "Register new user in the database",
        "operationId": "createUser",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Request success"
          },
          "400": {
            "description": "Invalid parameters - Message for invalid input parameters",
            "schema": { "$ref": "#/definitions/ApiResponse" }
          },
          "422": {
            "description": "Unprocessable Request - User already registered in the database",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "500": {
            "description": "Internal error - Message for request processing failure for undetectable generic errors",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "message for error description or detailed API response",
          "description": "Informational message with details of the request processing failure"
        }
      }
    }
  },
...
}